--- src/pages/sy.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import ExpoSalesPeriodSet from 'components/productExpoSalesPeriod/ExpoSalesPeriodSet';
-import PreOrderReservetionSet from 'components/productPreOrderReservation/PreOrderReservetionSet';
-
-const SY = () => {
-  return (
-    <div>
-      <ExpoSalesPeriodSet />
-      <PreOrderReservetionSet />
-    </div>
-  );
-};
-
-export default SY;
